Add basic support for table captions
In Pandoc, table captions (enabled by the `table_caption` option) can be
entire paragraphs. This is a little difficult to parse, because we need
to detect lines that might start a block quote, a code fence, or a
heading, since we support parsing these without a preceding blank line.
In the case of a code fence, we need to actually read the whole code
fence to see if it is valid. Therefore, this commit implements only
single-line captions.

Progresses #32